Revert "added few database models and crud"
This reverts commit ba6fa1881ec6feac5b0d3e8787014aad4c0bacd0.

+++ backend/tamalesHr/src/main/java/com/sjsu/cmpe272/tamales/tamalesHr/repository/EmployeeRepository.java
@@ -3,4 +3,7 @@ package com.sjsu.cmpe272.tamales.tamalesHr.repository;
 import com.sjsu.cmpe272.tamales.tamalesHr.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {}
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+
+
+}
